VERSION_E, _X

+++ build.sh
@@ -63,7 +63,27 @@ export F18_POS=1
 export F18_RNAL=1
 scripts/build_gz_push.sh Ubuntu_i686
 
+F18_VER=`cat VERSION`
+if [ ! -f /vagrant/F18_Ubuntu_i686_${F18_VER}.gz ] ; then
+   scripts/build_gz_push.sh Ubuntu_i686 ${F18_VER}
+   cp VERSION F18_VER
+   cp F18_VER /vagrant/
+   cp F18_Ubuntu_i686_${F18_VER}.gz /vagrant/
+fi
 
-cp VERSION F18_VER
-cp F18_VER /vagrant/
-cp F18_Ubuntu_i686_`cat F18_VER`.gz /vagrant/
+F18_VER_E=`cat VERSION_E`
+if [ -f VERSION_E ] &&  [ ! -f /vagrant/F18_Ubuntu_i686_${F18_VER_E}.gz ] ; then
+   scripts/build_gz_push.sh Ubuntu_i686 ${F18_VER_E}
+   cp VERSION_E F18_VER_E
+   cp F18_VER_E /vagrant/
+   cp F18_Ubuntu_i686_${F18_VER_E}.gz /vagrant/
+fi
+
+F18_VER_X=`cat VERSION_X`
+if [ -f VERSION_X ] &&  [ ! -f /vagrant/F18_Ubuntu_i686_${F18_VER_X}.gz ] ; then
+   scripts/build_gz_push.sh Ubuntu_i686 ${F18_VER_X}
+   cp VERSION_X F18_VER_X
+   cp F18_VER_X /vagrant/
+   cp F18_Ubuntu_i686_${F18_VER_X}.gz /vagrant/
+fi
+
